+++ R/zzz.R
@@ -1,7 +1,7 @@
 #' @importFrom Rdpack reprompt
 
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("Parallel numerical derivatives v.0.0.9 (2025-03-04).")
+  packageStartupMessage("Parallel numerical derivatives v.0.0.10 (2025-04-02).")
 
   # The number of cores is auto-detected based on the OS
   os <- Sys.info()[["sysname"]]