+++ NEWS.md
@@ -3,6 +3,7 @@
 - BUG: Matching in the Hessian is too slow -- de-duplicate first
 - BUG: 1x1 Hessians?
 - UX: make the warnings once-per-session; print the first error in runParallel in Grad
+- UX: add a pretty print method for the gradient, Jacobian, and step size
 - FEATURE: add an algorithm where the two lines intersect; should fix `sin'(pi/2)` with two branches
 - FEATURE: Implement de-duplication in the grid of points; reuse f0 and other precomputed values
 - FEATURE: If `h` is a character in `Grad`, extract the gradient directly if the order is 2